Revert "Change dtype of mask"
This reverts commit 5edcbb3cbf5e02182867b88acebb6a05f1e82a8b.

+++ demo.py
@@ -41,7 +41,7 @@ if __name__ == '__main__':
         masks.mul_(255).squeeze_(dim=1)
 
         # 이미지와 정답 정보를 GPU로 복사
-        images, masks = images.to(device), masks.type(torch.ByteTensor)
+        images, masks = images.to(device), masks.type(torch.LongTensor)
 
         # 예측
         with torch.no_grad():
